Add config settings for local testing
The use of site.url in templates and links to assets meant that, during
local testing, the local copy was not linking properly to itself.
Loading _local_config.yml after _config.yml with the --config serve
option means that localhost is used as site.url.

This file should be used in future to change configuration during
local testing.